Move platform and interface stuff in their own files in public directory
