+++ public/auto-string.h
@@ -23,6 +23,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <amtl/am-string.h>
+#include <amtl/am-algorithm.h>
 
 namespace ke {
