+++ public/reapi/mod_rehlds_api.cpp
@@ -1,5 +1,4 @@
 
-#include <amxmodx.h>
 #include "mod_rehlds_api.h"
 
 IRehldsApi*          RehldsApi;
